feat: accepted scope and rejected claim tracking
interactionCheck default now checks that requested scope does not extend
previously extended scopes

interaction results can now be passed an array of strings under
`consent.rejectedClaims` for claims that the end-user has rejected to
share. These claims will not be returned in either userinfo or id tokens

TODO:
- accepted claim tracking (resolved from scopes, claims parameter minus
the rejected ones) and interactionCheck that will request interaction if
claims are not accepted yet.